4.20-21

+++ part4/blogilista/controllers/blogs.js
@@ -4,14 +4,6 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
-
 blogsRouter.get('/', async (request, response, next) => {
     try {
         const blogs = await Blog.find({})
@@ -24,9 +16,7 @@ blogsRouter.get('/', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
     try {
-        const token = getTokenFrom(request)
-
-        const decodedToken = jwt.verify(token, process.env.SECRET)
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
         if (!token || !decodedToken.id) {
             return response.status(401).json({ error: 'token missing or invalid' })
         }
@@ -54,7 +44,21 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-        const blog = await Blog.findByIdAndRemove(request.params.id)
+        const token = getTokenFrom(request)
+
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(request.params.id)
+        console.log(user)
+        console.log(blog.user)
+        if (user._id === blog.user._id) {
+            await Blog.findByIdAndRemove(request.params.id)
+        } else {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
         response.status(204).end()
     } catch (exception) {
         next(exception)
